// src/types/generated/ConversationRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ConversationMessage } from "./ConversationMessage";
import type { ConversationSummary } from "./ConversationSummary";

export type ConversationRecord = { summary: ConversationSummary, messages: Array<ConversationMessage>, };